update to courses

+++ project_app/views.py
@@ -17,6 +17,7 @@ class Login(View):
             request.session["name"] = request.POST['name']
             return redirect('/SupHome/')
 
+
 class SupHome(View):
     def get(self, request):
         return render(request, "sup_home.html", {})
@@ -28,7 +29,8 @@ class SupAccounts(View):
         return render(request, "sup_accounts.html", {"roles": Roles.choices, "accounts": accounts})
 
     def post(self, request):
-        message = createAccount(request.POST['username'], request.POST['password'], request.POST['email'], request.POST['role'])
+        message = createAccount(request.POST['username'], request.POST['password'], request.POST['email'],
+                                request.POST['role'])
         accounts = list(User.objects.all())
         return render(request, "sup_accounts.html", {"roles": Roles.choices, "accounts": accounts, "message": message})
 
@@ -38,8 +40,10 @@ class RegisterCourses(View):
         return render(request, "register_courses.html")
 
     def post(self, request):
-        message = createCourse(request.POST['cor_id'], request.POST['cor_name'], request.POST['cor_sched'], request.POST['cor_cred'])
+        message = createCourse(request.POST['cor_id'], request.POST['cor_name'], request.POST['cor_sched'],
+                               request.POST['cor_cred'])
         return render(request, "register_courses.html", {"message": message})
+
 
 class SupEmail(View):
     def get(self, request):
@@ -50,19 +54,20 @@ class Account(View):
     def get(self, request):
         return render(request, "account.html", {})
 
+
 class SupCourses(View):
     def get(self, request):
         courses = list(Course.objects.all())
-        return render(request, "sup_courses.html", {"courses":courses})
-
-    def post(self, request):
-        Course.objects.get(courseid=request.POST["delete_course"]).delete()
-        courses = list(Course.objects.all())
         return render(request, "sup_courses.html", {"courses": courses})
 
+    def post(self, request):
+        if request.POST.get('delete_course'):
+            Course.objects.get(courseid=request.POST['delete_course']).delete()
+            courses = list(Course.objects.all())
+            return render(request, "sup_courses.html", {"courses": courses})
+        else:
+            return redirect('/RegisterCourses/')
+
 class CreateLab(View):
-    def get(self,request):
+    def get(self, request):
         return render(request, "sup_create_lab.html")
-
-
-
